// src/api_gen/models/ActivateUserDto.ts
/**
 * MOEPP CITES API
 * MOEPP CITES API docs
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ActivateUserDto {
    'status'?: ActivateUserDtoStatusEnum;
    'statusReason'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "ActivateUserDtoStatusEnum",
            "format": ""
        },
        {
            "name": "statusReason",
            "baseName": "statusReason",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ActivateUserDto.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ActivateUserDtoStatusEnum {
    Inactive = 'inactive',
    Verified = 'verified',
    Pending = 'pending',
    Approved = 'approved',
    Rejected = 'rejected',
    Expired = 'expired'
}