// src/api_gen/models/AuthRegisterLoginDto.ts
/**
 * MOEPP CITES API
 * MOEPP CITES API docs
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class AuthRegisterLoginDto {
    'email': string;
    'password': string;
    'language': AuthRegisterLoginDtoLanguageEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "email",
            "baseName": "email",
            "type": "string",
            "format": ""
        },
        {
            "name": "password",
            "baseName": "password",
            "type": "string",
            "format": ""
        },
        {
            "name": "language",
            "baseName": "language",
            "type": "AuthRegisterLoginDtoLanguageEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AuthRegisterLoginDto.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum AuthRegisterLoginDtoLanguageEnum {
    En = 'en',
    Mk = 'mk'
}