// src/api_gen/models/AuthUpdateDto.ts
/**
 * MOEPP CITES API
 * MOEPP CITES API docs
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class AuthUpdateDto {
    'password'?: string;
    'oldPassword'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "password",
            "baseName": "password",
            "type": "string",
            "format": ""
        },
        {
            "name": "oldPassword",
            "baseName": "oldPassword",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AuthUpdateDto.attributeTypeMap;
    }

    public constructor() {
    }
}