// src/api_gen/models/Channel.ts
/**
 * MOEPP CITES API
 * MOEPP CITES API docs
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Guild } from '../models/Guild';
import { HttpFile } from '../http/http';

export class Channel {
    'createdAt': Date;
    'updatedAt': Date | null;
    'deletedAt': Date | null;
    'id': number;
    'uid': string;
    'guild': Guild;
    'guildId': number;
    'type': ChannelTypeEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "createdAt",
            "baseName": "createdAt",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "updatedAt",
            "baseName": "updatedAt",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "deletedAt",
            "baseName": "deletedAt",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "uid",
            "baseName": "uid",
            "type": "string",
            "format": ""
        },
        {
            "name": "guild",
            "baseName": "guild",
            "type": "Guild",
            "format": ""
        },
        {
            "name": "guildId",
            "baseName": "guildId",
            "type": "number",
            "format": ""
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "ChannelTypeEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Channel.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ChannelTypeEnum {
    Construction = 'construction',
    Research = 'research',
    Training = 'training'
}