// src/api_gen/models/CreateChannelDto.ts
/**
 * MOEPP CITES API
 * MOEPP CITES API docs
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateChannelDto {
    'uid': string;
    'guildUid': string;
    'type': CreateChannelDtoTypeEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "uid",
            "baseName": "uid",
            "type": "string",
            "format": ""
        },
        {
            "name": "guildUid",
            "baseName": "guildUid",
            "type": "string",
            "format": ""
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "CreateChannelDtoTypeEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateChannelDto.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum CreateChannelDtoTypeEnum {
    Construction = 'construction',
    Research = 'research',
    Training = 'training'
}